+++ lib/object/GameCharacter.ts
@@ -1,0 +1,3 @@
+import { GameObject } from './generic/GameObject';
+
+export class GameCharacter extends GameObject {}