+++ lib/object/generic/interface/IGameCharacter.ts
@@ -1,0 +1,4 @@
+export interface IGameCharacter {
+  getHealth(): number;
+  setHealth(health: number): number;
+}