+++ lib/object/generic/interface/IGameState.ts
@@ -1,0 +1,4 @@
+export interface IGameState {
+  set(key: string, value: string | number | boolean | any): void;
+  get(key: string): string | number | boolean | any;
+}