+++ src/constant.js
@@ -1,1 +1,2 @@
 const CANVAS = document.querySelector('canvas');
+const DEBUG = true;