+++ src/v1/objects/state.object.js
@@ -1,0 +1,6 @@
+class State {
+  static Paused = 0;
+  static Running = 1;
+  static Stopped = 2;
+  static GameOver = 3;
+}